requirements.txt
requests
feedparser
beautifulsoup4
pyyaml
python-dateutil
